fixed users tg_id bug

+++ core/models/Users.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Integer
+from sqlalchemy import Integer,BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from .Base import Base
 from .Favorites import favorites_table
 
 
 class User(Base):
-    tg_id: Mapped[int] = mapped_column(Integer, unique=True, index=True, nullable=False)
+    tg_id: Mapped[BigInteger] = mapped_column(Integer, unique=True, index=True, nullable=False)
     username: Mapped[str]
     balance = mapped_column(Integer, default=0)
     favorites: Mapped[list["Product"]] = relationship(
